backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pymongo==4.6.0
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4
smtplib2==0.2.1